requirements.txt
flask==0.12.2
flask-restful==0.3.6
PyYAML==3.12
